Crate tardigrade_rt

Tardigrade runtime library.

The runtime provides a WorkflowManager in which workflows defined in a WASM module can be executed and persisted / restored. Interaction with a workflow (e.g., submitting messages to channel receivers or taking messages from senders) can be performed using WorkflowHandles.

Components

The crate defines two major abstractions necessary to run workflows:

Engine that provides a way to define, instantiate, run and snapshot workflows

Storage that provides a mechanism to durably persist the state of workflows and channels connecting them and the external world

For each of abstractions, there is a default implementation available: the wasmtime-powered engine and the in-memory, in-process storage (it still provides a way to (de)serialize a storage snapshot, so it’s not completely useless).

See the linked module docs for more details on abstractions.

Crate features

test

(Off by default)

Provides the test module with helpers for integration testing of workflows.

async-io

(Off by default)

Implements Schedule trait, necessary to instantiate async workflow handles, using the async-io crate.

Examples

Instantiating workflow

use tardigrade_rt::{
    engine::Wasmtime, manager::WorkflowManager, storage::LocalStorage,
};
use tardigrade::spawn::CreateWorkflow;

let module_bytes: Vec<u8> = // e.g., take from a file
let engine = Wasmtime::default();
let module = engine.create_module(module_bytes)?;
// It is possible to inspect module definitions:
for (workflow_name, interface) in module.interfaces() {
    println!("{workflow_name}: {interface:?}");
}

// Let's instantiate a manager and add the module to it.
let storage = LocalStorage::default();
let mut manager = WorkflowManager::builder(engine, storage)
    .build()
    .await?;
manager.insert_module("test", module).await;

// Workflows are created within a manager that is responsible
// for their persistence and managing channels, time, and child workflows.
let spawner = manager.spawner();
let builder = spawner.new_workflow::<()>("test::Workflow")?;
let (handles, self_handles) = builder.handles(|_| {}).await;
let new_workflow =
    builder.build(b"data".to_vec(), handles).await?;

// Let's initialize the workflow.
let receipt = manager.tick().await?.into_inner()?;
// `receipt` contains information about WASM code execution. E.g.,
// this will print the executed functions and a list of important
// events for each of executions:
println!("{:?}", receipt.executions());
// It's possible to communicate with the workflow using `self_handles`
// and/or `new_workflow`.

See WorkflowManager docs for examples of what to do with workflows after instantiation.

Persisting and restoring workflow

let manager: WorkflowManager<Wasmtime, (), LocalStorage> = // ...
let workflow = manager.storage().workflow(workflow_id).await.unwrap();
let persisted: &PersistedWorkflow = workflow.persisted();
// The persisted workflow can be serialized:
let json = serde_json::to_string(persisted)?;

// In case of using `LocalStorage`, the entire state can be serialized
// as well:
let mut storage = manager.into_storage();
let json = serde_json::to_string(&storage.snapshot())?;

Modules

engine

Workflow engine abstraction.

handle

Handles for workflows in a WorkflowManager and their components (e.g., channels).

manager

WorkflowManager and tightly related types.

receipt

Receipts for workflows and associated types.

storage

Async, transactional storage abstraction for storing workflows and channel state.

testtest

Helpers for workflow integration testing.

Structs

AsyncIoSchedulerasync-io

Scheduler implementation from async-io (a part of async-std suite).

Channels

Information about channels for a particular workflow interface.

ChildWorkflow

State of child workflow as viewed by its parent.

PersistedWorkflow

Persisted version of a workflow containing the state of its external dependencies (channels and timers), and its linear WASM memory.

ReceiverState

State of a workflow channel receiver.

SenderState

State of a workflow channel sender.

TaskState

State of a task within a workflow.

TimerState

State of a workflow timer.

Traits

Clock

Wall clock.

Schedule

Scheduler that allows creating futures completing at the specified timestamp.

Type Definitions

TimerFuture

Future for Schedule::create_timer().