Module tardigrade_rt::engine

Workflow engine abstraction.

Overview

A WorkflowEngine is responsible for:

Transforming a WASM module into one or more named workflow definitions

Instantiating zero or more workflow instances from a definition.

Running and persisting workflow instances

The crate provides a default engine implementation powered by the wasmtime crate.

Structs

ChildActions

Handle allowing to manipulate an (initialized) child workflow.

ReceiverActions

Handle allowing to manipulate a workflow channel receiver.

SenderActions

Handle allowing to manipulate a workflow channel sender.

TaskActions

Handle allowing to manipulate a workflow task.

TimerActions

Handle allowing to manipulate a workflow timer.

Wasmtime

WorkflowEngine based on the wasmtime crate.

WasmtimeDefinition

Workflow definition powered by the wasmtime crate.

WasmtimeInstance

Workflow instance powered by the wasmtime crate.

WasmtimeModule

Workflow module wrapping a [Module] from the wasmtime crate.

WorkflowData

Data associated with a workflow instance.

WorkflowPoll

Polling context for workflows.

Enums

ReportedErrorKind

Kinds of errors reported by workflows.

Traits

AsWorkflowData

Provides access to wrapped WorkflowData.

CreateWaker

Creating wakers in workflows. This is used in WorkflowPoll to unwrap the contained Poll result.

DefineWorkflow

Workflow definition.

PersistWorkflow

Workflow persistence logic.

RunWorkflow

Workflow instance execution logic.

WorkflowEngine

Workflow engine.

WorkflowModule

Workflow module.