Trait tardigrade_rt::engine::RunWorkflow

pub trait RunWorkflow: AsWorkflowData {
    fn create_main_task(&mut self, raw_args: &[u8]) -> Result<TaskId>;
    fn poll_task(&mut self, task_id: TaskId) -> Result<Poll<()>>;
    fn drop_task(&mut self, task_id: TaskId);
    fn wake_waker(&mut self, waker_id: WakerId);
    fn initialize_child(
        &mut self,
        stub_id: WorkflowId,
        result: Result<WorkflowId, HostError>
    );
    fn initialize_channel(&mut self, stub_id: ChannelId, channel_id: ChannelId);
}

Workflow instance execution logic.

A workflow instance wraps WorkflowData and can use its methods to manipulate the workflow in response to calls in the workflow logic (e.g., polling a next item for a channel receiver). Thus, WorkflowData methods conceptually map to WASM module imports. Conversely, methods in the RunWorkflow trait conceptually map to WASM module exports.

Required Methods

fn create_main_task(&mut self, raw_args: &[u8]) -> Result<TaskId>

Creates the main workflow task and returns its ID.

Errors

Returns an error if creating the main task fails.

fn poll_task(&mut self, task_id: TaskId) -> Result<Poll<()>>

Polls a task with the specified ID.

Errors

Returns an error if polling the task fails, e.g., due to a WASM trap.

fn drop_task(&mut self, task_id: TaskId)

Drops a task with the specified ID.

fn wake_waker(&mut self, waker_id: WakerId)

Wakes a waker with the specified ID.

fn initialize_child(
&mut self,
stub_id: WorkflowId,
result: Result<WorkflowId, HostError>
)

Notifies that the child workflow with the specified local_id, which was previously supplied to WorkflowData::create_workflow_stub(), has been initialized with the specified result. Either the child was allocated the ID wrapped in Ok(_), or an error has occurred during initialization.

fn initialize_channel(&mut self, stub_id: ChannelId, channel_id: ChannelId)

Notifies that the channel with the specified local_id, which was previously supplied to WorkflowData::create_channel_stub(), has been created and was allocated the provided channel_id.