Struct tardigrade_rt::handle::ErroredWorkflowHandle

pub struct ErroredWorkflowHandle<S> { /* private fields */ }

Handle for an errored workflow in a Runtime.

The handle can be used to get information about the error, inspect potentially bogus inbound messages, and repair the workflow. See the runtime docs for more information about workflow lifecycle.

Examples

let storage: StorageRef<'_, LocalStorage> = // ...
let workflow_id: WorkflowId = // ...
let workflow = storage.any_workflow(workflow_id).await.unwrap();
let workflow = workflow.unwrap_errored();
// Let's inspect the execution error.
let panic_info = workflow.error().panic_info().unwrap();
println!("{panic_info}");

// Let's inspect potentially bogus inbound messages:
let mut dropped_messages = false;
for message in workflow.messages() {
    let payload = message.receive().await?.decode()?;
    if is_bogus(&payload) {
        message.drop_for_workflow().await?;
        dropped_messages = true;
    }
}

if dropped_messages {
    // Consider that the workflow is repaired:
    workflow.consider_repaired().await?;
} else {
    // We didn't drop any messages. Assume that we can't do anything
    // and terminate the workflow.
    workflow.abort().await?;
}

Implementations

impl<S: Storage> ErroredWorkflowHandle<S>

pub fn id(&self) -> WorkflowId

Returns the ID of this workflow.

pub fn record(&self) -> &WorkflowRecord<()>

Returns a workflow record corresponding to this workflow, with the state field scrubbed.

pub fn error(&self) -> &ExecutionError

Returns the workflow execution error.

pub async fn abort(self) -> Result<(), ConcurrencyError>

Aborts this workflow.

Errors

Returns an error if abort fails because of a concurrent edit (e.g., the workflow is already aborted).

pub async fn consider_repaired(self) -> Result<(), ConcurrencyError>

Considers this workflow repaired, usually after dropping bogus messages for the workflow or otherwise eliminating the error cause.

Errors

Returns an error if repair fails because of a concurrent edit (e.g., the workflow is already repaired or aborted).

pub fn messages(&self) -> impl Iterator<Item = ErroneousMessage<&S>> + '_

Iterates over messages the ingestion of which may have led to the execution error.