Struct tardigrade_rt::handle::WorkflowHandle

pub struct WorkflowHandle<W, S> { /* private fields */ }

Handle to an active workflow in a Runtime.

A workflow handle allows getting handles for the channels specified in the workflow interface. The returned handles allow interacting with the workflow by sending messages and/or receiving messages.

Examples

use tardigrade::handle::{ReceiverAt, SenderAt, WithIndexing};
use tardigrade_rt::handle::WorkflowHandle;

// Assume we have a dynamically typed workflow:
let workflow: WorkflowHandle<(), _> = // ...
// We can create a handle to manipulate the workflow.
let handle = workflow.handle().await.with_indexing();

// Let's send a message via a channel.
let message = b"hello".to_vec();
handle[ReceiverAt("commands")].send(message).await?;

// Let's then take outbound messages from a certain channel:
let message = handle[SenderAt("events")]
    .receive_message(0)
    .await?;
let message: Vec<u8> = message.decode()?;

// It is possible to access the underlying workflow state:
let persisted = workflow.persisted().common();
println!("{:?}", persisted.tasks().collect::<Vec<_>>());
let now = persisted.current_time();

Implementations

impl<S: Storage> WorkflowHandle<(), S>

pub fn downcast<W: GetInterface>( self ) -> Result<WorkflowHandle<W, S>, AccessError>

Attempts to downcast this handle to a specific workflow interface.

Errors

Returns an error on workflow interface mismatch.

impl<W: WithHandle, S: Storage> WorkflowHandle<W, S>

pub fn id(&self) -> WorkflowId

Returns the ID of this workflow.

pub fn record(&self) -> &WorkflowRecord<()>

Returns a workflow record corresponding to this workflow, with the state field scrubbed.

pub fn interface(&self) -> &Interface

Returns the workflow interface.

pub fn persisted(&self) -> &PersistedWorkflow

Returns the snapshot of the persisted state of this workflow.

The snapshot is taken when the handle is created and can be updated using Self::update().

pub async fn update(&mut self) -> Result<(), ConcurrencyError>

Updates the snapshot contained in this handle.

Errors

Returns an error if the workflow was terminated.

pub async fn abort(self) -> Result<(), ConcurrencyError>

Aborts this workflow.

Errors

Returns an error if abort fails because of a concurrent edit (e.g., the workflow is already aborted).

impl<'a, W: WithHandle, S: Storage> WorkflowHandle<W, &'a S>

pub async fn handle(&self) -> HostHandles<'a, W, S>

Returns a handle for the workflow that allows interacting with its channels.