Type Alias tardigrade_rt::handle::RawMessageReceiver

pub type RawMessageReceiver<S> = MessageReceiver<Vec<u8>, Raw, S>;

MessageReceiver for raw bytes.

Aliased Type

struct RawMessageReceiver<S> { /* private fields */ }

Implementations

impl<T, C: Codec<T>, S: Storage> MessageReceiver<T, C, S>

pub fn channel_id(&self) -> ChannelId

Returns the ID of the channel this receiver is connected to.

pub fn channel_info(&self) -> &ChannelRecord

Returns the snapshot of the state of this channel.

The snapshot is taken when the handle is created and can be updated using Self::update().

pub async fn update(&mut self)

Updates the snapshot of the state of this channel.

pub async fn receive_message( &self, index: u64 ) -> Result<ReceivedMessage<T, C>, MessageError>

Takes a message with the specified index from the channel.

Errors

Returns an error if the message is not available.

pub fn receive_messages( &self, indices: impl RangeBounds<u64> ) -> impl Stream<Item = ReceivedMessage<T, C>> + '_

Receives the messages with the specified indices.

Since the messages can be truncated or not exist, it is not guaranteed that the range of indices for returned messages is precisely the requested one. It is guaranteed that indices of returned messages are sequential and are in the requested range.

pub fn can_manipulate(&self) -> bool

Checks whether this receiver can be used to manipulate the channel (e.g., close it). This is possible if the channel receiver is not held by a workflow.

This check is based on the local snapshot of the channel state and thus can be outdated.

pub async fn truncate(&self, min_index: u64)

Truncates this channel so that its minimum message index is no less than min_index. If Self::can_manipulate() returns false, this is a no-op.

pub async fn close(self)

Closes this channel from the host side. If the receiver is not held by the host, this is a no-op.

impl<T, C: Codec<T>, S: Storage + Clone> MessageReceiver<T, C, &S>

pub fn into_owned(self) -> MessageReceiver<T, C, S>

Converts the storage referenced by this receiver to an owned form.

impl<T: 'static, C: Codec<T>, S: StreamMessages> MessageReceiver<T, C, S>

pub fn stream_messages(&self, indices: RangeFrom<u64>) -> MessageStream<T, C>

Streams messages from this channel starting from the specified index.