Type Alias tardigrade_rt::handle::RawMessageSender

pub type RawMessageSender<S> = MessageSender<Vec<u8>, Raw, S>;

MessageSender for raw bytes.

Aliased Type

struct RawMessageSender<S> { /* private fields */ }

Implementations

impl<'a, T, C: Codec<T>, S: Storage> MessageSender<T, C, S>

pub fn channel_id(&self) -> ChannelId

Returns the ID of the channel this sender is connected to.

pub fn channel_info(&self) -> &ChannelRecord

Returns the snapshot of the state of this channel.

The snapshot is taken when the handle is created and can be updated using Self::update().

pub async fn update(&mut self)

Updates the snapshot of the state of this channel.

pub fn can_manipulate(&self) -> bool

Checks whether this sender can be used to manipulate the channel (e.g., close it) Self::channel_info() (i.e., this check can be outdated). This is possible if the channel sender is held by the host.

pub async fn send(&self, message: T) -> Result<(), SendError>

Sends a message over the channel.

Errors

Returns an error if the channel is full or closed.

pub async fn send_all( &self, messages: impl IntoIterator<Item = T> ) -> Result<(), SendError>

Sends the provided messages over the channel in a single transaction.

Errors

Returns an error if the channel is full or closed.

pub async fn close(self)

Closes this channel from the host side.

impl<T, C: Codec<T>, S: Storage + Clone> MessageSender<T, C, &S>

pub fn into_owned(self) -> MessageSender<T, C, S>

Converts the storage referenced by this sender to an owned form.