Module tardigrade_rt::manager

WorkflowManager and tightly related types.

See WorkflowManager docs for an overview and examples of usage.

Structs

DriveConfig

Configuration for driving workflow execution in a WorkflowManager.

ManagerSpawner

Specialized handle to a WorkflowManager allowing to spawn new workflows.

TickResult

Result of ticking a WorkflowManager.

WorkflowManager

Manager for workflow modules, workflows and channels.

WorkflowManagerBuilder

Builder for a WorkflowManager.

WouldBlock

An error returned by WorkflowManager::tick() if the manager cannot progress.

Enums

Termination

Terminal status of driving a WorkflowManager.

Traits

AsManager

Trait encapsulating all type params of a WorkflowManager.

Clock

Wall clock.

Schedule

Scheduler that allows creating futures completing at the specified timestamp.

Type Definitions

TimerFuture

Future for Schedule::create_timer().