Module tardigrade_rt::receipt

Receipts for workflows and associated types.

The root type is Receipt that records zero or more Executions. Each Execution records Events that have occurred when executing a specific function from a certain workflow definition. Events can relate either to a message channel connected to the workflow (ChannelEvent), or to a resource managed by the runtime (ResourceEvent), such as a task or a timer.

Structs

ChannelEvent

Event related to a workflow channel receiver or sender.

Execution

Execution of a top-level WASM function with a list of events that have occurred during the execution.

ExecutionError

Error occurring during workflow execution.

PanicInfo

Information about a panic in the workflow code.

Receipt

Receipt for executing tasks in a workflow.

ResourceEvent

Event related to a host-managed resource (a task or a timer).

StubEvent

Event related to a channel or workflow stub.

Enums

ChannelEventKind

Kind of a ChannelEvent.

Event

Event included into a Receipt.

ExecutedFunction

Executed top-level WASM function.

ResourceEventKind

Kind of a ResourceEvent.

ResourceId

ID of a host-managed resource.

StubEventKind

Kind of a StubEvent

StubId

ID of a workflow-managed stub resource.

WakeUpCause

Cause of waking up a workflow task.