Module tardigrade_rt::runtime

Runtime and tightly related types.

See Runtime docs for an overview and examples of usage.

Structs

DriveConfig

Configuration for driving workflow execution in a Runtime.

Runtime

Runtime for workflow modules, workflows and channels.

RuntimeBuilder

Builder for a Runtime.

RuntimeSpawner

Specialized handle to a Runtime allowing to spawn new workflows.

TickResult

Result of ticking a Runtime.

WouldBlock

An error returned by Runtime::tick() if the runtime cannot progress.

Enums

Termination

Terminal status of driving a Runtime.

WorkflowTickError

Errors that can occur when ticking a specific workflow.

Traits

AsRuntime

Trait encapsulating all type params of a Runtime.

Clock

Wall clock.

MapFormat

Type mapper from channel handles to a HandleFormat. Used as a boundary for RuntimeSpawner.

Schedule

Scheduler that allows creating futures completing at the specified timestamp.

Type Aliases

TimerFuture

Future for Schedule::create_timer().