Struct tardigrade_rt::runtime::Runtime

pub struct Runtime<E: WorkflowEngine, C, S> { /* private fields */ }

Runtime for workflow modules, workflows and channels.

A workflow runtime is responsible for managing the state and interfacing with workflows and channels connected to the workflows. In particular, a runtime supports the following operations:

Spawning new workflows (including from the workflow code) using a RuntimeSpawner handle obtained via Self::spawner()

Manipulating channels (writing / reading messages) and workflows using a StorageRef handle obtained via Self::storage()

Driving the contained workflows to completion, either manually or using a driver if the underlying storage supports event streaming

Workflow lifecycle

A workflow can be in one of the following states:

Active. This is the initial state, provided that the workflow successfully initialized. In this state, the workflow can execute, receive and send messages etc.

Completed. This is the terminal state after the workflow completes as a result of its execution or is aborted.

Errored. A workflow reaches this state after it panics. The panicking execution is reverted, but this is usually not enough to fix the error; since workflows are largely deterministic, if nothing changes, a repeated execution would most probably lead to the same panic. An errored workflow cannot execute until it is repaired.

The only way to repair a workflow so far is to make the workflow skip the message(s) leading to the error.

Examples

// A runtime is instantiated using the builder pattern:
let storage = LocalStorage::default();
let runtime = Runtime::builder(Wasmtime::default(), storage)
    .with_clock(AsyncIoScheduler)
    .build();
// After this, modules may be added:
let module: WasmtimeModule = // ...
runtime.insert_module("test", module).await;

// After that, new workflows can be spawned using `ManageWorkflowsExt`
// trait from the `tardigrade` crate:
use tardigrade::spawn::CreateWorkflow;

let spawner = runtime.spawner();
let definition_id = "test::Workflow";
// ^ The definition ID is the ID of the module and the name of a workflow
//   within the module separated by `::`.
let args = b"test_args".to_vec();
let builder = spawner.new_workflow::<()>(definition_id).await?;
let (handles, self_handles) = builder.handles(|_| {}).await;
let mut workflow = builder.build(args, handles).await?;
// Do something with `workflow`, e.g., write something to its channels
// using `self_handles`...

// Initialize the workflow:
let receipt = runtime.tick().await?.into_inner()?;
println!("{receipt:?}");
// Check the workflow state:
workflow.update().await?;
assert!(workflow.persisted().is_initialized());

Implementations

impl<E: WorkflowEngine, C: Clock, S: Storage> Runtime<E, C, S>

pub async fn tick(&self) -> Result<TickResult, WouldBlock>

Attempts to advance a single workflow within this runtime.

A workflow can be advanced if it has outstanding wakers, e.g., ones produced by expiring timers or by flushing outbound messages during previous workflow executions. Alternatively, there is an inbound message that can be consumed by the workflow.

Errors

Returns an error if the runtime cannot be progressed.

pub async fn tick_workflow( &self, workflow_id: WorkflowId ) -> Result<TickResult, WorkflowTickError>

Attempts to advance a specific workflow within this runtime.

Errors

Returns an error in any of the following cases:

The workflow with the specified ID does not exist.

The workflow is not currently active.

The workflow cannot be advanced (see WouldBlock)

impl<E: WorkflowEngine, S: Storage> Runtime<E, (), S>

pub fn builder(engine: E, storage: S) -> RuntimeBuilder<E, (), S>

Creates a builder that will use the specified storage.

impl<E: WorkflowEngine, C: Clock, S: Storage> Runtime<E, C, S>

pub async fn insert_module(&self, id: &str, module: E::Module) -> ModuleRecord

Inserts the specified module into the runtime.

Panics

Panics if the module id contains ineligible chars, such as :.

pub fn engine(&self) -> &E

Returns a reference to the execution engine.

pub fn storage(&self) -> StorageRef<'_, S>

Returns a reference to the underlying storage.

pub fn into_storage(self) -> S

Returns the encapsulated storage.

pub async fn in_transaction( &mut self ) -> Runtime<E, C, TransactionAsStorage<S::Transaction<'_>>>

Creates a storage transaction and uses it for the further operations on the runtime (e.g., creating new workflows / channels). Beware of TransactionAsStorage drawbacks documented in the type docs.

The original transaction can be extracted back using Self::into_storage() and then TransactionAsStorage::into_inner().

Examples

let storage = LocalStorage::default();
let mut runtime = Runtime::builder(Wasmtime::default(), storage)
    .build();
// Obtain a runtime operating in a transaction
let tx_runtime = runtime.in_transaction().await;
// Create a workflow and channels for it in a transaction
let spawner = tx_runtime.spawner();
let builder = spawner.new_workflow::<()>("test::Workflow").await?;
let (handles, _) = builder.handles(|_| {}).await;
builder.build(vec![], handles).await?;

// Unwrap the transaction and commit it.
if let Some(tx) = tx_runtime.into_storage().into_inner() {
    tx.commit().await;
}
// If the transaction is not committed, the workflow and all channels created
// for it are gone.

pub fn spawner(&self) -> RuntimeSpawner<'_, Self>

Returns a spawner handle that can be used to create new workflows.

pub fn raw_spawner(&self) -> RuntimeSpawner<'_, Self, Raw>

Returns a raw spawner handle that can be used to create new workflows. In contrast to Self::spawner(), the returned spawner uses channel IDs rather than channel handles to create workflows, and returns the ID of a created workflow rather than its handle.

pub async fn set_current_time(&self, time: DateTime<Utc>)

Sets the current time for this runtime. This may expire timers in some of the contained workflows.

impl<E, C, S> Runtime<E, C, Streaming<S>>where E: WorkflowEngine, C: Schedule, S: Storage + Clone,

pub async fn drive( self, commits_rx: &mut CommitStream, config: DriveConfig ) -> Termination

Drives this runtime using the provided config.