Module tardigrade_rt::storage

Async, transactional storage abstraction for storing workflows and channel state.

Overview

The core trait of this module is Storage, which produces storage transactions:

ReadonlyStorageTransaction for reading data

StorageTransaction for reading and writing data

Transactions must have at least snapshot isolation.

The data operated by transactions is as follows:

Entity type	Read trait	Write trait

ModuleRecord	ReadModules	WriteModules

ChannelRecord	ReadChannels	WriteChannels

WorkflowRecord	ReadWorkflows	WriteWorkflows

WorkflowWakerRecord	n/a	WriteWorkflowWakers

The ReadonlyStorageTransaction trait encompasses all read traits, and StorageTransaction encompasses both read and write traits. To simplify readonly transaction implementation without sacrificing type safety, there is Readonly wrapper.

In terms of domain-driven development, modules, channels and workflows are independent aggregate roots, while workflow wakers are tied to a workflow. _Record types do not precisely correspond to the relational data model, but can be straightforwardly mapped onto one. In particular, WorkflowRecord has the workflow state as a type param. The state is modelled via enum dispatch on the WorkflowState enum and its variants (e.g., ActiveWorkflowState for active workflows).

Besides the Storage trait, the module provides its local in-memory implementation: LocalStorage. It has LocalTransaction transactions and provides LocalStorageSnapshot as a way to get (de)serializable snapshot of the storage.

See also

StorageRef, a wrapper around a Storage that allows accessing workflow and channel sender / receiver handles

Structs

ActiveWorkflowState

State of an active workflow.

ChannelRecord

State of a workflow channel.

CommitStream

Stream of commits returned from Streaming::stream_commits().

CompletedWorkflowState

State of a completed workflow.

DefinitionRecord

Storage record for a workflow definition.

ErroneousMessageRef

Reference for a potentially erroneous message in ErroredWorkflowState.

ErroredWorkflowState

State of an errored workflow.

LocalReadonlyTransaction

Readonly transaction used by LocalStorage.

LocalStorage

Local in-memory Storage.

LocalStorageSnapshot

Serializable snapshot of a LocalStorage.

LocalTransaction

Transaction used by LocalStorage.

MessageEvent

Event signalling that one or more messages were added to the channel.

ModuleRecord

Storage record for a WorkflowModule.

ModuleRecordMut

Thin wrapper around ModuleRecord that allows accessing its fields and changing only module bytes.

Readonly

Wrapper for StorageTransaction implementations that allows safely using them as readonly transactions.

Streaming

Storage that streams events about new messages when they are committed.

StreamingTransaction

Wrapper around a transaction that streams MessageEvents on commit. Used as the transaction type for Streaming storages.

WorkflowRecord

State of a workflow.

WorkflowWakerRecord

Record associating a WorkflowWaker with a particular workflow.

Enums

MessageError

Error retrieving a message from a workflow channel. Returned by ReadChannels::channel_message().

MessageOrEof

Message from a channel or the channel end marker.

WorkflowSelectionCriteria

Workflow selection criteria used in insert_waker_for_matching_workflows().

WorkflowState

State of a WorkflowRecord.

WorkflowWaker

Waker for a workflow.

Traits

ReadChannels

Allows reading stored information about workflow channels.

ReadModules

Allows reading stored information about WorkflowModules.

ReadWorkflows

Allows reading information about workflows.

ReadonlyStorageTransaction

Storage transaction with readonly access to the storage.

Storage

Async, transactional storage for workflows and workflow channels.

StorageTransaction

Storage transaction with read/write access to the storage.

StreamMessages

Allows streaming messages from a certain channel as they are written to it.

WriteChannels

Allows modifying stored information about channels.

WriteModules

Allows modifying stored information about WorkflowModules.

WriteWorkflowWakers

Allows modifying stored information about WorkflowWaker.

WriteWorkflows

Allows modifying stored information about workflows.