Trait tardigrade_rt::storage::ReadChannels

pub trait ReadChannels {
    fn channel<'life0, 'async_trait>(
        &'life0 self,
        id: ChannelId
    ) -> Pin<Box<dyn Future<Output = Option<ChannelRecord>> + Send + 'async_trait>>
   where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn channel_message<'life0, 'async_trait>(
        &'life0 self,
        id: ChannelId,
        index: usize
    ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>, MessageError>> + Send + 'async_trait>>
   where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn channel_messages(
        &self,
        id: ChannelId,
        indices: RangeInclusive<usize>
    ) -> BoxStream<'_, (usize, Vec<u8>)>;
}

Allows reading stored information about workflow channels.

Required Methods

fn channel<'life0, 'async_trait>(
&'life0 self,
id: ChannelId
) -> Pin<Box<dyn Future<Output = Option<ChannelRecord>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Retrieves information about the channel with the specified ID.

fn channel_message<'life0, 'async_trait>(
&'life0 self,
id: ChannelId,
index: usize
) -> Pin<Box<dyn Future<Output = Result<Vec<u8>, MessageError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Gets a message with the specified 0-based index from the specified channel.

Errors

Returns an error if the message cannot be retrieved.

fn channel_messages(
&self,
id: ChannelId,
indices: RangeInclusive<usize>
) -> BoxStream<'_, (usize, Vec<u8>)>

Gets messages with indices in the specified range from the specified channel.