Trait tardigrade_rt::storage::ReadWorkflows

pub trait ReadWorkflows {
    fn count_active_workflows<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = usize> + Send + 'async_trait>>
   where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn workflow<'life0, 'async_trait>(
        &'life0 self,
        id: WorkflowId
    ) -> Pin<Box<dyn Future<Output = Option<WorkflowRecord>> + Send + 'async_trait>>
   where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn nearest_timer_expiration<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Option<DateTime<Utc>>> + Send + 'async_trait>>
   where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Allows reading information about workflows.

Required Methods

fn count_active_workflows<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = usize> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Returns the number of active workflows.

fn workflow<'life0, 'async_trait>(
&'life0 self,
id: WorkflowId
) -> Pin<Box<dyn Future<Output = Option<WorkflowRecord>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Retrieves a snapshot of the workflow with the specified ID.

fn nearest_timer_expiration<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Option<DateTime<Utc>>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Finds the nearest timer expiration in all active workflows.