Trait tardigrade_rt::storage::StorageTransaction

pub trait StorageTransaction: Send + Sync + WriteModules + WriteChannels + WriteWorkflows + WriteWorkflowWakers + WorkerStorageConnection<Error = Infallible> {
    // Required method
    fn commit<'async_trait>(
        self
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait;
}

Storage transaction with read/write access to the storage.

Transactions must satisfy the ACID semantics. In particular, they must apply atomically and be isolated (i.e., not visible to other transactions until committed).

Self::commit() must be called for before the transaction is dropped. Explicit transaction rollback support is not required; all transactions instantiated by a Runtime are guaranteed to eventually be committed (save for corner cases, e.g., panicking when the transaction is active). If rollback is supported by the storage, it is assumed to be the default behavior on transaction drop. It would be an error to commit transactions on drop, since this would produce errors in the aforementioned corner cases.

Required Methods

fn commit<'async_trait>( self ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: 'async_trait,

Commits this transaction to the storage. This method must be called to (atomically) apply transaction changes.