Trait tardigrade_rt::storage::WriteChannels

pub trait WriteChannels: ReadChannels {
    fn allocate_channel_id<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = ChannelId> + Send + 'async_trait>>
   where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn insert_channel<'life0, 'async_trait>(
        &'life0 mut self,
        id: ChannelId,
        state: ChannelRecord
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
   where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn manipulate_channel<'life0, 'async_trait, F>(
        &'life0 mut self,
        id: ChannelId,
        action: F
    ) -> Pin<Box<dyn Future<Output = ChannelRecord> + Send + 'async_trait>>
   where
        F: 'async_trait + FnOnce(&mut ChannelRecord) + Send,
        'life0: 'async_trait,
        Self: 'async_trait;
    fn push_messages<'life0, 'async_trait>(
        &'life0 mut self,
        id: ChannelId,
        messages: Vec<Vec<u8>>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
   where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn truncate_channel<'life0, 'async_trait>(
        &'life0 mut self,
        id: ChannelId,
        min_index: usize
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
   where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Allows modifying stored information about channels.

Required Methods

fn allocate_channel_id<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = ChannelId> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Allocates a new unique ID for a channel.

fn insert_channel<'life0, 'async_trait>(
&'life0 mut self,
id: ChannelId,
state: ChannelRecord
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Creates a new channel with the provided state. If the channel with the specified ID already exists, does nothing (i.e., the channel state is not updated in any way).

fn manipulate_channel<'life0, 'async_trait, F>(
&'life0 mut self,
id: ChannelId,
action: F
) -> Pin<Box<dyn Future<Output = ChannelRecord> + Send + 'async_trait>>where
F: 'async_trait + FnOnce(&mut ChannelRecord) + Send,
'life0: 'async_trait,
Self: 'async_trait,

Changes the channel state and selects the updated state.

fn push_messages<'life0, 'async_trait>(
&'life0 mut self,
id: ChannelId,
messages: Vec<Vec<u8>>
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Pushes one or more messages into the channel.

fn truncate_channel<'life0, 'async_trait>(
&'life0 mut self,
id: ChannelId,
min_index: usize
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Truncates the channel so that min_index is the minimum retained index.