Trait tardigrade_rt::storage::WriteModules

pub trait WriteModules: ReadModules {
    fn insert_module<'life0, 'async_trait>(
        &'life0 mut self,
        module: ModuleRecord
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
   where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn update_tracing_metadata<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        module_id: &'life1 str,
        metadata: PersistedMetadata
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
   where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
}

Allows modifying stored information about WorkflowModules.

Required Methods

fn insert_module<'life0, 'async_trait>(
&'life0 mut self,
module: ModuleRecord
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Inserts the module into the storage.

fn update_tracing_metadata<'life0, 'life1, 'async_trait>(
&'life0 mut self,
module_id: &'life1 str,
metadata: PersistedMetadata
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Updates tracing metadata for the module.

It is acceptable to merge the provided metadata JSON to the current one to prevent edit conflicts.