Trait tardigrade_rt::storage::WriteWorkflows

pub trait WriteWorkflows: ReadWorkflows {
    fn allocate_workflow_id<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = WorkflowId> + Send + 'async_trait>>
   where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn insert_workflow<'life0, 'async_trait>(
        &'life0 mut self,
        state: WorkflowRecord
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
   where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn workflow_for_update<'life0, 'async_trait>(
        &'life0 mut self,
        id: WorkflowId
    ) -> Pin<Box<dyn Future<Output = Option<WorkflowRecord>> + Send + 'async_trait>>
   where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn update_workflow<'life0, 'async_trait>(
        &'life0 mut self,
        id: WorkflowId,
        state: WorkflowState
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
   where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn workflow_with_wakers_for_update<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Option<WorkflowRecord<ActiveWorkflowState>>> + Send + 'async_trait>>
   where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn workflow_with_consumable_channel_for_update<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Option<WorkflowRecord<ActiveWorkflowState>>> + Send + 'async_trait>>
   where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Allows modifying stored information about workflows.

Required Methods

fn allocate_workflow_id<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = WorkflowId> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Allocates a new unique ID for a workflow.

fn insert_workflow<'life0, 'async_trait>(
&'life0 mut self,
state: WorkflowRecord
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Inserts a new workflow into the storage.

fn workflow_for_update<'life0, 'async_trait>(
&'life0 mut self,
id: WorkflowId
) -> Pin<Box<dyn Future<Output = Option<WorkflowRecord>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Returns a workflow record for the specified ID for update.

fn update_workflow<'life0, 'async_trait>(
&'life0 mut self,
id: WorkflowId,
state: WorkflowState
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Updates the state of a workflow with the specified ID.

fn workflow_with_wakers_for_update<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Option<WorkflowRecord<ActiveWorkflowState>>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Finds an active workflow with wakers and selects it for update.

fn workflow_with_consumable_channel_for_update<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Option<WorkflowRecord<ActiveWorkflowState>>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Finds an active workflow with an inbound channel that has a message or EOF marker consumable by the workflow, and selects the workflow for update.