Struct tardigrade_rt::test::MockScheduler

pub struct MockScheduler { /* private fields */ }

Available on crate feature test only.

Mock wall clock and scheduler.

Examples

A primary use case is to use the scheduler with a WorkflowManager for integration testing:

// We need `Streaming` storage to drive a `WorkflowManager`.
let storage = Arc::new(LocalStorage::default());
let (mut storage, storage_task) = Streaming::new(storage);
let mut commits_rx = storage.stream_commits();
task::spawn(storage_task);

// Set the mocked wall clock for the workflow manager.
let scheduler = MockScheduler::default();
let mut manager = WorkflowManager::builder(Wasmtime::default(), storage)
    .with_clock(scheduler.clone())
    .build()
    .await?;
let manager = Arc::new(manager); // to simplify handle management

let inputs: Vec<u8> = // ...
let spawner = manager.spawner();
let builder = spawner.new_workflow::<()>("test::Workflow")?;
let (handles, self_handles) = builder.handles(|_| {}).await;
builder.build(inputs, handles).await?;

// Spin up the driver to execute the `workflow`.
let mut self_handles = self_handles.with_indexing();
let events_rx = self_handles.remove(SenderAt("events")).unwrap();
let mut events_rx = events_rx.stream_messages(0..);

let manager = manager.clone();
task::spawn(async move {
    manager.drive(&mut commits_rx, DriveConfig::new()).await
});

// Advance mocked wall clock.
let now = scheduler.now();
scheduler.set_now(now + chrono::Duration::seconds(1));
// This can lead to the workflow progressing, e.g., by emitting messages
let message = events_rx.next().await.unwrap();
// Assert on `message`...

Implementations

impl MockScheduler

pub fn with_expirations() -> (Self, impl Stream<Item = DateTime<Utc>> + Unpin)

Creates a mock scheduler together with a stream that notifies the consumer about new timer expirations.

pub fn next_timer_expiration(&self) -> Option<DateTime<Utc>>

Returns the expiration for the nearest timer, or None if there are no active timers.

pub fn now(&self) -> DateTime<Utc>

Returns the current timestamp.

pub fn set_now(&self, now: DateTime<Utc>)

Sets the current timestamp for the scheduler.