Struct tardigrade_rt::engine::WorkflowData

pub struct WorkflowData { /* private fields */ }

Data associated with a workflow instance.

Implementations

impl WorkflowData

Channel-related functionality.

pub fn create_channel_stub(&mut self, stub_id: ChannelId) -> Result<()>

Starts creating a channel. When the channel initialization is complete, it will be reported via initialize_channel().

As of right now, channel initialization cannot fail, nor can it be cancelled by the workflow logic.

Arguments

stub_id must be unique across all invocations of create_channel_stub() for a single workflow. E.g., it can be implemented as an incrementing static.

Errors

Returns an error if a channel cannot be created.

pub fn receiver(&mut self, id: ChannelId) -> ReceiverActions<'_>

Returns an action handle for the receiver with the specified ID.

Panics

Panics if the receiver with id is not owned by the workflow. Note that receivers for the channel with ID 0 (the “pre-closed” channel) are not considered to be owned by any workflow, despite a workflow potentially acquiring receiver(s) for this channel on initialization.

pub fn sender(&mut self, id: ChannelId) -> SenderActions<'_>

Returns an action handle for the sender with the specified ID.

Panics

Panics if a sender with id is not owned by the workflow. Note that senders for the channel with ID 0 (the “pre-closed” channel) are not considered to be owned by any workflow, despite a workflow potentially acquiring sender(s) for this channel on initialization.

impl WorkflowData

Spawning-related functionality.

pub fn request_definition( &mut self, stub_id: u64, definition_id: &str ) -> Result<()>

Initializes access to a workflow definition with the specified ID. When the initialization is complete, it will be reported via resolve_definition().

Errors

Returns an error if the workflow has no capability to access workflow definitions.

pub fn create_workflow_stub( &mut self, stub_id: WorkflowId, definition_id: &str, args: Vec<u8>, channels: HandleMap<ChannelId> ) -> Result<()>

Starts initializing a child workflow with the specified parameters. When the child initialization is complete, it will be reported via initialize_child().

As of right now, child initialization cannot be cancelled by the workflow logic.

Arguments

stub_id must be unique across all invocations of create_workflow_stub() for a single workflow. E.g., it can be implemented as an incrementing static.

Errors

Returns an error if the workflow has no capability to spawn child workflows.

pub fn child(&mut self, id: WorkflowId) -> ChildActions<'_>

Returns an action handle for the child workflow with the specified ID.

Panics

Panics if the child with id does not exist in the workflow.

impl WorkflowData

Task-related functionality.

pub fn spawn_task(&mut self, task_id: TaskId, task_name: String) -> Result<()>

Records a spawned task.

pub fn task(&mut self, id: TaskId) -> TaskActions<'_>

Returns an action handle for the task with the specified ID.

Panics

Panics if the task with id does not exist in the workflow.

impl WorkflowData

Time-related functionality.

pub fn current_timestamp(&self) -> DateTime<Utc>

Returns the current timestamp.

pub fn create_timer(&mut self, definition: TimerDefinition) -> TimerId

Creates a new timer.

pub fn timer(&mut self, id: TimerId) -> TimerActions<'_>

Returns an action handle for the timer with the specified ID.

Panics

Panics if the timer with id does not exist in the workflow.

impl WorkflowData

pub fn persisted(&self) -> &PersistedWorkflowData

Returns persisted data for the encapsulated workflow.

pub fn report_error_or_panic( &mut self, error_kind: ReportedErrorKind, message: Option<String>, filename: Option<String>, line: u32, column: u32 )

Reports an error or panic that has occurred in the workflow.

pub fn send_trace(&mut self, trace: TracingEvent)

Reports a tracing event that has occurred in the workflow.