Module tardigrade_rt::handle

Handles for workflows in a Runtime and their components (e.g., channels).

The root of the handle hierarchy is StorageRef, a thin wrapper around a Storage. Its methods allow accessing workflows and channels contained in the storage.

Examples

let storage: StorageRef<_> = // ...
let workflow = storage.workflow(1).await.unwrap();
// Workflow channels can be accessed via the `workflow` handle:
let handle = workflow.handle().await.with_indexing();
handle[ReceiverAt("commands")].send(b"test".to_vec()).await?;
let message = handle[SenderAt("events")].receive_message(0).await?;

// Channel handles also can be accessed directly:
let sender = storage.sender(1).await.unwrap();
let sender = MessageSender::<u64, Json, _>::try_from_raw(sender)?;
sender.send_all([42, 555]).await?;

Structs

CompletedWorkflowHandle

Handle to a completed workflow in a Runtime.

ConcurrencyError

Concurrency errors for modifying operations on workflows.

ErroneousMessage

Handle for a potentially erroneous message consumed by an errored workflow.

ErroredWorkflowHandle

Handle for an errored workflow in a Runtime.

MessageReceiver

Handle for an workflow channel Sender that allows taking messages from the channel.

MessageSender

Handle for a workflow channel Receiver that allows sending messages via the channel.

ReceivedMessage

Message received from a workflow channel.

StorageRef

Thin wrapper around a Storage that provides methods for accessing and manipulating channels and workflows.

WorkflowHandle

Handle to an active workflow in a Runtime.

Enums

AnyWorkflowHandle

Handle to a workflow in an unknown state.

Type Aliases

HostHandles

Host handles of a shape specified by a workflow Interface and provided by a Storage.

RawMessageReceiver

MessageReceiver for raw bytes.

RawMessageSender

MessageSender for raw bytes.