Struct tardigrade_rt::storage::LocalStorage

pub struct LocalStorage { /* private fields */ }

Local in-memory Storage.

LocalStorage can be serialized by taking a snapshot and then restored from it. Since workflow module bytes constitute the largest serialization part, the snapshot provides a method to manipulate them, e.g. in order to store module bytes separately.

Examples

let engine = Wasmtime::default();
let mut storage = LocalStorage::default();
// Remove messages consumed by workflows.
storage.truncate_workflow_messages();
let runtime = Runtime::builder(engine, storage).build();
// Do something with the runtime...

let mut storage = runtime.into_storage();
let mut snapshot = storage.snapshot();
for mut module in snapshot.modules_mut() {
    // Save modules to the file system, rather than using the storage.
    let filename = format!("{}.wasm", module.id);
    fs::write(&filename, &module.bytes).await?;
    module.set_bytes(filename.into_bytes());
}
let serialized = serde_json::to_string_pretty(&snapshot)?;

// Restoring the storage:
let mut snapshot: LocalStorageSnapshot<'_> =
    serde_json::from_str(&serialized)?;
for mut module in snapshot.modules_mut() {
    let filename = str::from_utf8(&module.bytes)?;
    module.set_bytes(fs::read(filename).await?);
}
let storage = LocalStorage::from(snapshot);

Implementations

impl LocalStorage

pub fn snapshot(&mut self) -> LocalStorageSnapshot<'_>

Returns a snapshot of the storage. The returned snapshot can be (de)serialized with serde.

pub fn truncate_workflow_messages(&mut self)

Automatically truncates messages received by workflows.