Trait tardigrade_rt::storage::ReadWorkflows

pub trait ReadWorkflows {
    // Required methods
    fn count_active_workflows<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = u64> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn workflow<'life0, 'async_trait>(
        &'life0 self,
        id: WorkflowId
    ) -> Pin<Box<dyn Future<Output = Option<WorkflowRecord>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn nearest_timer_expiration<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Option<DateTime<Utc>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Allows reading information about workflows.

Required Methods

fn count_active_workflows<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = u64> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Returns the number of active workflows.

fn workflow<'life0, 'async_trait>( &'life0 self, id: WorkflowId ) -> Pin<Box<dyn Future<Output = Option<WorkflowRecord>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Retrieves a snapshot of the workflow with the specified ID.

fn nearest_timer_expiration<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Option<DateTime<Utc>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Finds the nearest timer expiration in all active workflows.